Fix formatting check in CI (#146)
The `.prettierignore` was ignoring some Typescript files that we'd
actually want to check. Rather than have CI check against everything
(`prettier .`) with a list of what to exclude, remove the exclusion list
and just check all the Typescript files in `src`.

See https://github.com/carolynzech/kani-vscode-extension/pull/2 for
evidence that it works.

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 and MIT licenses.