Update to use source based coverage instead of line based (#145)
Updates the extension to use source-based coverage instead of line-based
coverage. See #139 for screenshots of how it looks.

1.
[a2e37a5](https://github.com/model-checking/kani-vscode-extension/pull/145/commits/a2e37a589886cfec7173b2a3c5a365516b09bf83)
just applies the changes from #139 in a single commit, rebased on top of
main.
2.
[e2e3a40](https://github.com/model-checking/kani-vscode-extension/pull/145/commits/e2e3a402fb6ec066cc8fa040727f667ad413443e)
changes the extension to show the "Get coverage info" button
automatically. Before, our documentation instructed a user to [go to
settings to enable the
button](https://github.com/carolynzech/kani-vscode-extension/blob/25489e8ad813f965428608870426f8e7b2ed35c8/docs/user-guide.md?plain=1#L82).
I couldn't find the setting to do that, and I figured if I couldn't find
it, our users wouldn't be able to either. I also don't think it makes
sense to hide it behind a setting. We can print the warning that it's
unstable, but the green/red coverage visual is one of the strongest
arguments for using the Kani extension instead of the command line, so
we should have this feature front and center.
3.
[69f376b](https://github.com/model-checking/kani-vscode-extension/pull/145/commits/69f376b2104343e2f9200002d6c24c1f1daa5662)
just removes comments and unused code.
4.
[8a99568](https://github.com/model-checking/kani-vscode-extension/pull/145/commits/8a99568b3bb30bc965a8af62647bc22054eafe73)
updates the documentation to reflect the shift from line-based to
source-based coverage, and the removal of the toggle option from
[e2e3a40](https://github.com/model-checking/kani-vscode-extension/pull/145/commits/e2e3a402fb6ec066cc8fa040727f667ad413443e).
5.
[05d0246](https://github.com/model-checking/kani-vscode-extension/pull/145/commits/05d0246c5b1f70eecbf7200948e99260e055702a)
introduces more custom types, so that we use those instead of any types.